Small improvements to UX - Resolves #52: add `tools` keyword - Resolves #50: Add DD check on activation
