+++ src/commands/runPrompt.ts
@@ -14,51 +14,6 @@ import { randomUUID } from "crypto";
 
 type PromptOption = 'local-dir' | 'local-file' | 'remote';
 
-const START_DOCKER_COMMAND = {
-    'win32': 'Start-Process -NoNewWindow -Wait -FilePath "C:\\Program Files\\Docker\\Docker\\Docker Desktop.exe"',
-    'darwin': 'open -a Docker',
-    'linux': 'systemctl --user start docker-desktop',
-};
-
-const checkDockerDesktop = () => {
-
-    // Coerce the error to have an exit code
-    type DockerSpawnError = Error & { code: number };
-
-    try {
-        const res = spawnSync("docker", ["version"]);
-
-        if (res.error) {
-            // Using -1 to indicate docker is not installed
-            (res.error as DockerSpawnError).code = -1;
-            throw res.error;
-        }
-
-        if (res.status !== 0) {
-            const err = new Error(`Docker command exited with code ${res.status} and output the following error: ${res.error || res.stderr.toString()}`);
-            // Using -1 as a fallback, should have already been caught by res.error
-            (err as DockerSpawnError).code = res.status || -1;
-            throw err;
-        }
-
-        // @ts-expect-error
-    } catch (e: DockerSpawnError) {
-        const platform = process.platform;
-        const actionItems = e.code !== -1 ? [(platform in START_DOCKER_COMMAND ? "Start Docker" : "Try again")] : ["Install Docker Desktop", "Try again"];
-        return vscode.window.showErrorMessage("Error starting Docker", { modal: true, detail: (e as DockerSpawnError).toString() }, ...actionItems).then(async (value) => {
-            switch (value) {
-                case "Start Docker":
-                    spawnSync(START_DOCKER_COMMAND[platform as keyof typeof START_DOCKER_COMMAND], { shell: true });
-                case "Install Docker Desktop":
-                    vscode.env.openExternal(vscode.Uri.parse("https://www.docker.com/products/docker-desktop"));
-                    return;
-                case "Try again":
-                    return 'RETRY';
-            }
-        });
-    }
-};
-
 const getWorkspaceFolder = async () => {
     const workspaceFolders = vscode.workspace.workspaceFolders;
 
@@ -87,10 +42,6 @@ const getWorkspaceFolder = async () => {
 export const runPrompt: (secrets: vscode.SecretStorage, mode: PromptOption) => void = (secrets: vscode.SecretStorage, mode: PromptOption) => vscode.window.withProgress({ location: vscode.ProgressLocation.Window, cancellable: true }, async (progress, token) => {
     progress.report({ increment: 1, message: "Starting..." });
     postToBackendSocket({ event: 'eventLabsPromptRunPrepare', properties: { mode } });
-    const result = await checkDockerDesktop();
-    if (result === 'RETRY') {
-        return runPrompt(secrets, mode);
-    }
 
 
     progress.report({ increment: 5, message: "Checking for OpenAI key..." });
@@ -200,7 +151,7 @@ export const runPrompt: (secrets: vscode.SecretStorage, mode: PromptOption) => v
                     await writeToEditor(`${header} ROLE ${role}${content ? ` (${content})` : ''}\n\n`);
                     break;
                 case 'functions-done':
-                    await writeToEditor('\n```'+`\n\n*entering tool*\n\n`);
+                    await writeToEditor('\n```' + `\n\n*entering tool*\n\n`);
                     break;
                 case 'message':
                     await writeToEditor(json.params.content);
